+++ auth-common/src/main/java/net/therap/filter/UserAuthFilter.java
@@ -1,0 +1,8 @@
+package net.therap.filter;
+
+/**
+ * @author apurboturjo
+ * @since 8/3/25
+ */
+public class UserAuthFilter {
+}
