+++ authserver/src/main/java/net/therap/controller/PublicKeyController.java
@@ -1,0 +1,5 @@
+/**
+* @author apurboturjo
+* @since 8/3/25
+*/public class PublicKeyController {
+}
